SDC-14809: Sort jars in SDCClassLoader
I had a user who had two machines and our MapR components - worked on
one machine, but failed on the other one - eventually this turned out to
be SDC-14808, but getting there to understand it was quite cumbersome.

We're already changing the classpath order we get from the OS/FS/Java to
push our own jars to the front, so let's sort the rest of the classpath
too. It would not solve the problem here but it would make it consistent
and thus much easier to troubleshoot.

Change-Id: Ie468ff8e7986915aec59d9eba8b98855dde13d7a
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32308
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <Santhosh@streamsets.com>