SDC-14825: Do not run query validation if query validation is disabled in JDBC Origin
This patch disables actual query validation if the config option is set
to disable them. The reason for doing so is that we had a user who is
using the origin against non-supported database (Redshift) where the
validation query dumped the whole table - it seems that us setting hints
for fetch size and max rows did not really helped.

Since there is no database agnostic way to ask for no rows - even
the famous 'limit' isn't actually supported across all databases since
it's not ANSI SQL, we don't actually have many other options here.

I was contemplating the idea of simply pushing back but then the option
to disable query validation is already in for really long time and it
seems that us issuing the query validation even if the option is enabled
seems weird.

Change-Id: Ia8a08344e5e231ac980d8a232d752542e9eb729a
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32318
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>