SDC-14275: Preview with Max Batch Size (records) set to > 10 throws an error
Back in SDC-12710 we have introduced a new check to warn the user if
they configured batch size bigger then is allowed by data collector's
configuration. The intention is good as user had no way how to easily
confirm that.

The implementation sadly decided to change each and every origin which
is not only not-scalable, non-enforceable, the origin also doesn't have
the knowledge in which 'mode' it's running in all cases. This leads to
the error being generated in situations when it definitely should not. I
wish we would have implemented the check differently, but it's a bit
late for that.

There are at least two ways that I'm aware of where the origin will
issue the error incorrectly - 1) preview and 2) snapshot. The preview is
the most user visible as now pretty much every single preview leads to
this error which is very annoying. Luckily the origin does know whether
preview mode is on or off, so this patch changes all the origins the
original change affected and avoids the check in the preview mode. This
removes the unnecessary error for the user. Sadly we can't solve for 2)
as there is no way for the origin to know whether it's running in
snapshot mode or not.

Change-Id: I054ca35cf50af4b4029fa7c10bb561ca39277f01
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32448
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <Santhosh@streamsets.com>