SDC-14867. SFTP origin causes the pipeline to stop when reading files with no permission.
This patch adds a isReadable method to RemoteFile (STFP, FTP) to detect whether a file is readable. This is to alleviate the problems especially when using WHOLE_FILE where suddenly the desination cannot read the file. This could happen in the below scenario:
Permissions to list directory, but file does not have read permissions

The patch is trying to create a stream to the remote file, as there are no easy to integrate apis in both FTP/STFP interface. The SFTP Interface allows getting file attributes which contain the permission, though there is not easy way to get the current uid/gid and check for user/group/other permission read.FTP Interface has a isReadable covenience method but unfortunately the implementation always returns true, so this patch opts for creating a stream for the file as a way to make sure we actually can read the file.

Testing:
1. Manual testing with FTP and STTP environments.
2. STF tests

Change-Id: I372f5315a0ce2bf57653c56f0ca5d1c4a521de50
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32478
Reviewed-by: Robert Kanter <rkanter@streamsets.com>