SDC-14882. JDBC Query Consumer closing the connection after each batch
In SDC-13540 we changed the way the stage handles the database connection. The connection
is closed after a batch is generated and that implicitly closes the ResultSet. Any new
batch is built again with the ResultSet pointing to the first record in the table.

This is an issue when working in non-incremental mode. We must keep the ResultSet open
across batches, to iterate over new records when the queried table is longer than the max
batch size configured. Otherwise, the stage will indefinitely generate batches with the
first batch_size records.

Change-Id: I2a4ea472a9be89881134b41d77a656d4a3937f2b
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32523
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>