SDC-14895. AWS Credential Store permission errors during initialization
During credential store initialization (i.e. at SDC startup), we validate AWS
Secrets Manager connectivity by querying a dummy (non-existing) secret and
processing the exception.

We expected the exception to be a ResourceNotFoundException. Any other error is
propagated as a connection/configuration failure. However, if a user does not
have global permissions for the Describe and GetValue operations, the expected
exception should be an AccessDeniedException instead. This patch modifies the
validation to contemplate both cases.

Also two additional minor changes:
- Mini-refactor in alwaysRefresh option parsing.
- Catch all the AWS Secrets Manager exceptions when retrieving keys, so that all
  the related error messages have the AWS_SECRETS_MANAGER_CRED_STORE_03 error code.

Change-Id: I41cfcbc3fc0eb49790b1d040702efc7b98329ac0
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32590
Reviewed-by: Sebastian Sanchez <sebas@streamsets.com>