SDC-14842. Retrieve Connection configs from SCH
This patch allows an SDC pipeline to use a Connection by retrieving the
Connection's configs from SCH.  The ConnectionRetriever class takes care of that
and most "real" changes are in there, ConfigInjector.

To retrieve the Connection config, it was necessary to pipe the user who's
running the pipeline down to the ConnectionRetriever, which required a lot of
minor changes everywhere; similar to when we had to pipe down the blobstore.
The blobstore is currently not being used here anymore, but I left it in in case
we use it later.

TRANSFORM-2037 will do the equivalent for Transformer, but that will require
some additional work because of the way Transformer runs its pipelines.  That
may require additional SDC changes due to shared code, but that can be done with
a followup Jira.

At the moment, there is no way to select a Connection when building a pipeline.
In the meantime, it's possible to directly edit the pipeline's JSON (either in
the DB or by exporting it, editing it, and importing it) to add the Connection's
ID.

Change-Id: If3b43d947e112a93dddaf30d108617914c6b74c7
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32718
Reviewed-by: Alison Taylor <alison@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>