SDC-14788 [telemetry] Failures to report should not affect save frequency and should have limited backoff
Previously, there were issues with backoff on reporting failures
that would skip the minutely stats save and would retry at
unreasonably long intervals.

Now, we will never skip saving on the minutely timer, and will
retry with exponential backoff from 2 minutes to 24 hours. This
allows us to recover from a short-term availability or load issue
on the upload side in a matter of minutes, without overloading
the system with retries, and without waiting an unreasonably long
time to retry (previously it could wait for hundreds of millions
of days).

We also never turn off stats if there are reporting failures.

Testing:
* Unit test
* Manually ran SDC with shorter reporting and roll times, and an
  invalid URL to cause uploads to fail. Observed logs that showed
  regular minutely save and the expected rolls, and with a reporting
  back off that doubled as far as 64 minutes before I stopped it.

Change-Id: Ie407e3404b6220e9d6809477494f2c2dcd5b936d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32768
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <Santhosh@streamsets.com>