SDC-14879 Robust terminal transitions for SDC Pipeline State
Now only validate state transitions in Standalone and Cluster
Runner when it is the start of a user-initiated action. This lets
us prevent users from making illegal state changes, but without
problems where we know that the pipeline is in a certain state,
like we have finished the run method or connected and found the
remote cluster to have terminated the job, but some internal logic
error is preventing us from performing the transition. It is a bug
if we find ourselves in an illegal transition, but it is better to
set the state to what we know is true than to error out and keep
the pipeline marked as running or something.

As a safety precaution, we will never move from an inactive state
unless it is a valid transition, even for system-initiated
transitions.

Testing:
* Unit tests for both Standalone and Cluster Runners
* Started up SDC and made sure simple piplines started and stopped
* Started up CDH cluster with SDC, patched jar, ran YARN batches:
** Start and let it finish normally
** Start and stop it
** Start and kill it on the YARN side (via Cloudera Manager)

Change-Id: I18f10685dd0402cbee7dcb477395faf78e5f5f4c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32819
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>