SDC-14971: Missing configs are not added while validating Test Origin
You could argue that this is a bug in the JDBC Query origin that while
adding the Root Field config, we have not updated the upgrader to add
that config explicitly. But then we do have an explicit logic in our
validator that is adding missing configs with default values and I know
that bunch of states are depending on that behavior. Thus this patch
fixes the backend code where we obviously missed adding missing configs
to this additional stage type.

Change-Id: Ia43143509e368f8e9cd852daa998730329441ffa
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32839
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Martin Balzamo <martin@streamsets.com>