SDC-15016. Oracle CDC issues with PGA memory in Oracle 12
It was observed in Oracle 12.2.0.1 that LogMiner sessions are not implicitly
closed after starting a new LogMiner session with the same database
connection. This worked for Oracle 11, but it seems to be
implementation-dependent and consequently we cannot rely on it.

When CONTINUOUS_MINE option is enabled (i.e. for Oracle versions < 19), we now
explicitly close the LogMiner session after each mining window. Also, we restore
the behavior prior to SDC-13751: mining windows can span across the future, in
which case querying the LogMiner metadata returns a blocking cursor.

When CONTINUOUS_MINE option is disabled (i.e. Oracle 19c onwards), we still keep
the implementation introduced in SDC-13751, with mining windows bounded to the
current database time as LogMiner cursors are necesarilly non-blocking. PGA
memory is not leaked because the stage keeps the same LogMiner session between
mining windows and the redo log list for mining is manually updated.

Change-Id: If37214ce7e486c29e78eced709eda67d5c6869b6
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33018
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>