SDC-15051. SDC changes for TRANSFORM-2037
This patch contains SDC-side changes required by TRANSFORMER-2037 (i.e. to get
Connections running in Transformer).

While SDC can simply call SCH via REST to get the Connections, that won't work
with Transformer because it'll run the actual pipeline in Spark or elsewhere.
In a nutshell, Transformer will download the Connections used by a particular
pipeline, serialize them to a file, and pass that file to Spark the same way
it's already passing other objects (e.g. pipeline, offset, etc).

ConfigInjector had previously been written to make two passes: the first pass
looks for Connection IDs, and builds a map of Connection Type to Connection ID;
and the second pass actually fetches them and puts them in.  This is fine as
long as each Stage only has at most one of each Connection Type (which is true
as far as we have planned); however, if we need to serialize all of the
Connections used by a particular pipeline into a single file, we could easily
have conflicts if two different Stages were to use two different Connections of
the same Type.  On top of that, this was a weird way to do it.  Instead, we now
fetch the Connections in the first pass, and store a mapping of Connection ID to
Connection, so that the second pass can simply put them in.

The other important change is that the patch bubbles up the mapping of
Connection ID to Connection higher up in the call stack.  This has a few
benefits, including that it can be re-used.  It turns out that our existing code
rebuilds a pipeline multiple times during validation and other actions, so by
being able to pass the already fetched Connections back down, we can prevent extra
REST calls.  Another benefit is that Transformer (in TRANSFORMER-2037) can grab
the Connections that were fetched for a pipeline so it can serialize and pass
them to Spark.  The downside to this change is that it leaves
Connection-related code littered throughout a lot more of the codebase than we
originally had.

This patch also adds the ignoreUnknown annotation to ConnectionConfigurationJson
so that SCH doesn't fail to send the Connection configs to SDC/Transformer when
the UI puts extra stuff in there that Jackson doesn't like.

Change-Id: I536cf1affd662ad8c155a7a0152a331308604268
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33264
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Joan Sala <joan.sala@streamsets.com>