SDC-15099. Hide connection information on the dynamic preview pipeline
The connectionSelection attribute is added to the connection verifier. When a preview for a connection is requested, we will create a map with the connection configuration and pass it all the way down to the ConfigInjector to fill in the verifier stage. When building the verifier origin stage we won't set the connection configurations.

Basically we are delaying the injection of the connection configurations enough so that the values won't appear on the pipeline canvas for the user to see.

Change-Id: Ied05860597b69b5d82153c9a9c6955259c43a865
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33275
Reviewed-by: Robert Kanter <rkanter@streamsets.com>