SDC-15263. Protect Oracle CDC from race conditions during a log rotation
This race condition can happen when LogMiner's CONTINUOUS_MINE option is
disabled and Oracle CDC is managing the redo log list to mine (i.e. for Oracle
versions >= 19).

The problem arises under two circumstances, both related to transient states
of the online redo logs:

- The current online redo log is being rotated while the Oracle CDC mining
window reached the present time in database. During the transition to the new
current log, we must ensure the new current log is already open and ready.

- One of the online redo logs needed for the Oracle CDC mining window is shown
as archived, but the archiving process is still pending to finish (i.e. the
corresponding copy is not shown as ready).

This patch enables the stage to detect when a required redo log is in a
transient state and follow an sleep-and-retry strategy to solve it, rather than
fail with a StageException.

Changes in this patch:
- Detect transient states and propagate the error to the public
methods (LogMinerSession#start, LogMinerSession#preloadDictionary).

- Adapt OracleCDCSource#generateRecords loop: when that error is detected, sleep
thread and retry.

Change-Id: I04e9c560aec1be86a909edf29ce5716b33ee927b
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33666
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>