TRANSFORM-2120: Remove tabs for specific execution mode
This patch removes the EMR and DataProc tabs for Transformer - moving
all those configs to CLUSTER tab to be consistent with DataBricks and
other Transformer cluster types.

To help with Basic/Advanced, this patch also tags all SDC-EMR (yes there
is Transformer EMR and Data Collector EMR which are two different things
...) configs as ADVANCED since they do depend on ADVANCED config
(executionMode).

This cleans up the UI considerably both in Transformer as well as in
Basic Mode in Data Collector.

Change-Id: Ic260c49b97028173ae65674860a01885ed92c8fe
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33848
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>