SDC-15337: Allow stages to differentiate between empty and pre-filled default value for ListBeanModel
For a long time no matter what exactly was configured as a default value
of ListBeanModel, we always forced empty list in the extractor which UI
interpreted to "pre-create" first entry of the list.

Now with the Basic/Advanced, this is a problem since we need to make
sure that Advance lists are indeed empty (as opposed defaulting to size
one).

This patch changes our internal extraction code to separate those two
states and then updates bunch of default values to be in line with the
new distinction. If the defaultValue is empty string, UI wil pre-create
one-item list. If the defaultValue is explicitly set to empty array by
using "[]" we will default to empty list.

Change-Id: Ia982b4cea2e7410f91a0c68f580afaf6b57d7055
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33866
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>