SDC-15396: Increase default value of max batch size
We have checks around to detect if someone is creating batch larger
then what is allowed in SDC and we issue an error for that. The problem
is that when someone creates a snapshot we override the max batchSize
to 10 (hardcoded in the UI as far as I can tell) which always leads to
this error.

The stage itself can't say whether it's running in a batch mode or not
which is very annoying. Thus this patch proposes to increase the default
snapshot size to the same value as is default max batch size so that we
know that by default, we never issue the validation error.

Change-Id: I056eb16bd47b3c5e38115d64a3cfa872e83ebdb2
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33951
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>