SDC-15332. Oracle CDC preview does NPE
This patch fixes the NPE that is happening ORacle CDC origin. the reason being we are not setting offset bean when previewing and a NPE is thrown later when nextOffset = recordOffset.offset.toString() is called.

Given that preview does not save offset back, its safe to initialize the offset.

Testing:
1. STF

Change-Id: Id6b037c2600d60523fb1ac11e7ac48f0af4bcab4
Reviewed-on: https://review.streamsets.net/c/datacollector/+/33952
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>