SDC-15480. Fix stagelib Connections verifier issue
Removes the "library" from the various Connection objects and JSONs - we weren't
really use it anyway; and the Connection doesn't really care about it either.
When you run a pipeline, we're going to use whatever stagelib that stage is
using regardless.  Instead, the "library" field is now found in the Connection
Verifiers, where it can matter (e.g. db driver, Hadoop jars version, etc).  The
JSON that provides the various Connection types to the FE now includes an array
of "verifierDefinitions" instead of just one, so that we can (eventually) ask
the user to pick which stagelib they want to run the Verifier with.  For now,
DPM-8714 will hardcode it to the first one in the array.  In any case, the FE
will pass the selected Verifier Definition to the backend as part of the Dynamic
Preview Request (which it was already doing), and things will work pretty much
how they did before.

The ClassLoaderStageLibraryTask logic to load the Connections and Connection
Verifiers has been updated accordingly.  Now that there's no "library" in a
Connection, they're all identical and library-agnostic, so it doesn't matter
which one we load.  If there's a version conflict, we choose the oldest one to
be the most compatible, but we may need to revisit that in the future.  For the
Verifiers, we now load all of them; if there is more than one for the same type,
we keep them in a set now, instead of arbitrarily picking one.

Change-Id: I2691e7222afdd87ee0c3b454e20eb13fb2855c45
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34091
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Joan Sala <joan.sala@streamsets.com>