SDC-15530. cassandra testcontainers version is wrong in cassandra_3-lib.
org.testcontainers.cassandra version was changed due to unit tests failing from RELEASE (which is not a fixed version and we don't know when it will change) to 1.14.3 in SDC-15157

This patch  makes the same change to casandra_3 lib.

While running the tests, i have verified this is actually running against a CASANDRA 3 container, See the output from docker ps when test was running.

176ed9d63c63        cassandra:3.11.2                    "docker-entrypoint.s…"   6 seconds ago       Up
 4 seconds        0.0.0.0:32809->7000/tcp, 0.0.0.0:32808->7001/tcp, 0.0.0.0:32807->7199/tcp, 0.0.0.0:32
805->9042/tcp, 0.0.0.0:32804->9160/tcp   infallible_golick

Change-Id: I66b905618388aeba9e14e85ce934227c4996f958
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34244
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Eddie Park <eddie@streamsets.com>