SDC-15588. S3 Connection Verifier sporadically fails with UnknownHostException cause
Not exactly sure why (might be some weird JVM DNS caching issue?) but
occasionally we'll get an UnknownHostException on the S3 Connection Verifier,
even though everything is fine.  This patch basically wraps that check into a
loop with a backoff to let it try 5 times before giving up, with a delay of 3,
6, 9, and 12 seconds between tries; which'll be about 30 seconds.  That should
be more than enough time to verify if it's a real UnknownHostException or this
flakey non-real problem.  While not the most elegant solution, from my testing,
this does appear to be robust.

Change-Id: Idfa8dac82184a20ab682bf657bd03ac52fd6957e
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34276
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>