SDC-12124. Add check box to Delay Processor for "Skip Delay on Empty Batch"
Added the "Skip Delay on Empty Batch" property to the Delay processor configuration.
Updated the processor version and implemented the upgrader definition.

Manually tested: the new property appears on the UI;
when checked and all records are rejected the delay is skipped;
also tested upgrading from old pipeline to new version.
Integration tests will be implemented in a separate change set.

Change-Id: Ic75457d2a34f76c24522b1d996999eedf41e5cd8
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34453
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>