SDC-15685: Explicitly load MySQL JDBC drivers in MySQL BinLog
ServiceLoader that is used to register and load JDBC drivers is known to
be buggy on JDK8 when multiple classloaders are in use. We've provided
workarounds for those issues back in SDC 3.10 which is primarily this
code:

https://github.com/streamsets/datacollector/blob/master/jdbc-protolib/src/main/java/com/streamsets/pipeline/lib/jdbc/HikariPoolConfigBean.java#L343

Sadly the MySQL Bin Log is not using that code to ensure that known
drivers are properly loaded. I did not wanted to refactore the whole
library just to get that in, so instead I've added code that explicitly
checks for MySQL drivers.

Change-Id: I990c3df74e25d09a111dcb9e85a2b0a3d770bac7
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34469
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>