SDC-14852. LocalFS event isn't consistent
To make the behaviour consistent for all types of file systems includeing remote
file systems we need to exclude the schema. The event record contains the filepath
attribute, not the uri attribute, so we need to use only the file path.

A separate JIRA ticket was created to add the URI attribute to the event record.

There is separate change set with integration tests.

Change-Id: Ia0469aa235105909b3d73bdf28a98fbfc4551037
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34502
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>