SDC-15747. TestUDPSource#testPrivilegedPort in basic-lib does not apply to Mac OS as written
The testPrivilegedPort test in basic-lib assumes that a non-root user
will not be able to bind to a port number less than 1024. This is no
longer true on Mac OS starting 10.14. This change updates the test to
not expect a bind issue in case the operating system is Mac OS.

Change-Id: Ib396f2291eb814706530efba30af8244fbc19a04
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34722
Reviewed-by: Dima Spivak <dima@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>