SDC-15608. Validate query in ElasticSearch origin
Implemented a method to validate a query.
The method will send a standard validation request to an ElasticSearch server.
This will work to validate a broken query and for broken JSONs.
Tested manually against Elasticsearch_5.2.0, Elasticsearch_5.3.1, Elasticsearch_6.8.12, Elasticsearch_7.9.0
Also tested for the incremental mode since it requires a placeholder in a query.
Appropritate integration tests will be implemented separately.

Change-Id: I0398c5eb803f7eca283f9f87f5b0d10b3ebe1a86
Reviewed-on: https://review.streamsets.net/c/datacollector/+/34807
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>