SDC-15855. Don't resolve username and password if AWS security is used
When AWS security is selected we do not validate if the username and password are set,
but we still try getting the values. Getting the values itself may fail and block a user.
Here we add a check if the AWS security mode is enabled and if so, we bypass getting the values.

Change-Id: Ia207960f7de0ba1d648c9d65decd24a606977399
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35116
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>