SDC-15590. Elasticsearch origin for EL5 and EL6 does not support plus in topic name
It's allowed for index and mapping to have the + character in their names.
This character is also allowed in a URL path (it's not allowed for query param names only),
but in ES5/6 there is a bug which makes the character impossible in a URL path.
Thus for these versions of ES the character MUST be encoded as well.

In this change set we introduce PathEscape whose implementations are responsible for path encoding.
Depending on the ES version appropriate implementation of PathEscape is loaded
and is called to perform the encoding as required by an ES version.
We expect that at runtime only one implementation of PathEscape gets loaded.
The implementation loading is implemented using Java SPI.

The change set also contains unit tests to verify the loading works as expected.

Appropriate STF tests will be implemented in the scope of the following task https://issues.streamsets.com/browse/SDC-15861

Change-Id: I5d602991aa1c4e67902f1c8fe4b7160f0a23d9b8
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35130
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>