SDC-15872: Disconnect MySQL Bin Log client if if it's not connected
We're using a third-party library to read and parse MySQL's binlog. This library
spawns a "keep-alive" thread. We are of course disconnecting the library on stage
destroy() phase, but here is where a small nuance is causing us troubles. The precise
code on our side is:

if (client != null && client.isConnected()) {
  try {
    client.disconnect();
  } catch (IOException e) {
    LOG.warn("Error disconnecting from MySql", e);
  }
}

E.g. we check if the client exists (client != null) and whether it's connected and
if so, disconnect. That seems completely normal on a first read. Sadly what seems
to be happening is that when the library encounters the critical error - a second
slave with the id has connected - it will disconnect, but will not shut down the
keep-alive thread. So the method isConnector returns false, we never explicitly
call disconnect() and JVM will be forever running left-over thread that is
constantly re-connecting to MySQL and initiating a session for bin log consuming.

It seems that by removing the client.isConnected() all works properly - e.g. no
thread is left after the stage is done and it seems that it's safe to call
disconnect() even if isConnected() returns false.

Change-Id: I2b7a2f6f867dc4e820b9a6c3f564cf902cfbc27c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35204
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>