SDC-15576. Setting query interval to 0 in ElasticSearch might lead to Interrupt Exception
We investigated the original issue claiming an error happens when the query interval is
set to zero, but we were not able to reproduce it, nether the reporter was able to do
that. Although our investigation demonstrated there was another issue that might occur
when the query interval wasn't zero. In some rare circumstances the interrupted
exception may be thrown. This patch set addresses the issue we identified, also it fixes
a unit test and adds another one to test incremental mode with small query interval.

The cause of the issue is that
1. In the while loop of the isFinished method we need to check if the stage is
terminated and break if that is the case
2. If ThreadUtil.sleep returns true we do need to set the interrupt thread since the
thread wasn't interrupted. But if sleep returns false, that means the thread was
interrupted but the method cleared the interrupted flag, so we need to set it again.

In one of the unit tests the assumption was to stop the stage when two empty batches
are returned. That is not required, one empty batch should be the event for us to stop
the stage.

Change-Id: I0386739d3f3782816057490cbeb5f377d24102db
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35304
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>