SDC-15923: NullPointerException if using str:regExCapture in Pipeline parameters
It seems that we are calling incorrect method on the ELEvaluator object.
The ElEval from API have two methods - eval and evaluate. The difference
is actually important as the eval will create ThreadLocal variable that
EL functions can use to store values inside.

We haven't found it so far because majority of implicit ELs (those that
doesn't depend on record information) don't actually need variables. The
regExCapture is an exception though.

Change-Id: I8c919578892d73ce38402d8e2e972944d3e16c90
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35371
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>