SDC-15848: Jenkins flakyness in com.streamsets.datacollector.restapi
In SDC-15761 we have introduced a new field called "roleARN" into
AWSConfig. This field is both a CredentialValue as well as contains a
default value set. This forces the framework to create instance of
ClearCredentialValue class that is then passed down for the default
value.

This is actually not a problem in normal execution because we already
have JsonConfigurator for the REST layer that uses our own
ObjectMapperFactory that provides serialization for
ClearCredentialValue. The problem is just in the unit test that doesn't
use this  configurator and thus have problems to serialize the default
value.

Change-Id: Ic6089e73cbfc5a16b6b9ebd004ad5dc2aad5a04c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35382
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>