SDC-15937: Kinesis Consumer doesn't work properly
Few issues with this one - first of all there was a weirdness in the
upgrader logic. Java was handling all versions until 6 and then to
version 8 whereas YAML was handling version 7 and then and 9. That is
not supported by the backend code though - java must end at one version
where yaml picks it up - not to have few versions handled in java and
few in yaml. So this patch moves the problematical upgrade of 7 to Java
and voila all is working.

There was existing 6 to 7 upgrade method that actually wasn't written
properly so I had to update it too.

Change-Id: I8b51feffb6ffd4626a7205cb09612f34de18b15d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35418
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>