SDC-15948. Propagate initial mining window upper limit to OracleCDCSource#generateRecords
Before this patch the initial mining window upper limit was re-computed in
OracleCDCSource#generateRecords. When CONTINUOUS_MINE is disabled, this could
cause a missmatch between the actual mining window in use by the LogMiner
session and the upper limit re-computed in generateRecords that eventually might
provoke data loss. This patch fixes it by passing the complete mining window
limit to generateRecords.

Also includes a minor refactor for the OracleCDCSource#startLogMiner method.

Change-Id: I117afa61662b7e385f854c57a6a26c0d2f8d3ded
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35453
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Toni Martinez <toni@streamsets.com>