SDC-15968. Skip LogMiner dictionary preloading when source is online catalog
Oracle CDC is unnecessarily preloading a logminer dictionary from the redo logs
when the dictionary source is the online catalog. Skip that step, which also has
the inconvenient of forcing the user to create a logminer dictionary in the redo
logs even for Oracle CDC configurations that do not need it.

Change-Id: I62faf54fadda0e0eecd4189f173772d94401042a
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35512
Reviewed-by: Marina Pera <marina@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 7b7362d66946811c0bb78d6d51ee16b9bfa1405e)
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35340
Reviewed-by: Martin Balzamo <martin@streamsets.com>