SDC-15993: Do not lowercase column names in PostgreSQL Metadata processor by default
I believe that the lowercasing behavior was done before because the
stage wasn't properly enclosing object names -- which is unfortunate
"workaround". Since we do support enclosing now, we can get rid of the
workaround.

To keep compatibility with older pipelines, the lowercasing behavior is
exposed as a configuration on Advanced tab. It defaults to false (e.g.
do no lowercase) for new pipelines but older pipelines are upgraded to
True to preserve backward compatibility.

Change-Id: I25af2d722b4928c806f44f8edccd3d6a746eb174
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35578
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>