SDC-15919. Failed to get ctime: Directory spooler concatenates directory path twice
When a glob pattern is used we store absolute path in offsets,
it is also used as a key in the map of processed files.
The reason of the error is that when we reconstruct the full file name
we concatenate the absolute path in the offset with the directory from the
configuration property which contains the glob pattern. As a result of the
concatenation the path is invalid. This later when we compare files by date
causes the reported exception to be thrown. Because of another issue the
exception is ignored, no data duplication nor data loses also happen.

In this fix we just check if the path is absolute we use it as is without
changes.

This is a short term solution, a longer term solution to use paths relative
to the longest directory path which doesn't contain a glob pattern should be
implemented as specified here https://issues.streamsets.com/browse/SDC-16016

Change-Id: I15658a30c2b9543a32c0f45751e8244f4d087052
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35607
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>