SDC-16084: Do not remove Legacy Tab entries in JDBC stages
It seems that this is an example of StreamSets Cloud backport
that brought up more changes that it should have. Long story short,
in StreamSets Cloud we never had the "Legacy" tab, so the upgraders
were made to explicitly remove those two properties when importing
pipeline from Data Collector. When backporting the changes back
from StreamSets Cloud to Data Collector, we have not noticed that
those properties actually do exists in Data Collector and left the
upgraders intact. So they are essentially dropping those two
properties since 3.16.0.

Change-Id: I5a553e1ac55d68fd6f6c631caf7d492f80a07ab6
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35791
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>