SDC-16106. Fix SQS pipeline data loss on pipeline stop
In the previous implementation there was no distinction between cases
when a record reached a destination and cases when it didn't reach any
destination because the pipeline had been interrupted. But in the both
case the previous implementation committed a record (removed a message
from a queue).
With this fix we will support At least one and At most one modes,
for the At least one mode a message will not be removed from a queue
if a record doesn't reach a destination.

Change-Id: If66aede901cbe13423226c6862c41682e23ee6c8
Reviewed-on: https://review.streamsets.net/c/datacollector/+/35895
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>